requirements.txt
python-telegram-bot==20.7
wakeonlan==3.1.0
flask==2.3.3